+++ src/renderer/bindings/BindedAppBar.tsx
@@ -1,23 +1,19 @@
 import React from "react"
 import * as _ from 'lodash'
-import {bindWith} from './base'
-import ResourceStore from '../stores/resourceStore'
-import SettingStore from '../stores/settingStore'
-import SnippetStore from '../stores/snippetStore'
+import useResourceStr from '../stores/resource'
+import useSettingStr from '../stores/setting'
+import useSnippetStr from '../stores/snippet'
 import AppBar from '../components/AppBar'
+import {observer} from 'mobx-react-lite'
 
-export default bindWith(['resourceStore', 'snippetStore', 'settingStore'], (props: {
-    resourceStore: ResourceStore,
-    settingStore: SettingStore,
-    snippetStore: SnippetStore
-}) => {
-    let rs = props.resourceStore
-    let ss = props.settingStore
-    let snippet = props.snippetStore
+export default observer(() => {
+    let rs = useResourceStr()
+    let ss = useSettingStr()
+    let snippet = useSnippetStr()
     return <AppBar
         query={rs.query}
         onAddIconClicked={ss.showSettingModal.bind(ss)}
-        onCommentIconClicked={snippet.showSnippetModal.bind(snippet)}
+        onCommentIconClicked={snippet.newSnippet.bind(snippet)}
         onEditIconClicked={()=>{}}
         onQueryChange={rs.updateQuery.bind(rs)}
         onSubmit={rs.reload.bind(rs)}