+++ src/renderer/bindings/BindedConfirmAlert.tsx
@@ -1,13 +1,11 @@
 import React from "react"
 import * as _ from 'lodash'
-import {bindWith} from './base'
 import ConfirmAlert from '../components/ConfirmAlert'
-import {BasicStore} from "../stores/common"
+import useBasicStr from "../stores/basic"
+import { observer } from "mobx-react-lite"
 
-export default bindWith(['basicStore'], (props:{
-    basicStore: BasicStore
-}) => {
-    let bs = props.basicStore
+export default observer(() => {
+    let bs = useBasicStr()
     return <ConfirmAlert
         visible={bs.confirmAlertVisible}
         title={bs.confirmAlertTitle}