+++ src/renderer/stores/snippet.ts
@@ -1,13 +1,13 @@
 import { observable, action, flow } from 'mobx'
 import * as _ from 'lodash'
-import {invokeRPC, notify} from './common'
+import {createContext, useContext} from 'react'
+import {invokeRPC, notify} from './basic'
+import {store as rs} from './resource'
 
-export default class SnippetStore {
+export class SnippetStore {
     @observable visible: boolean = false
     @observable content: string = ""
     @observable id: string = ""
-
-    saved: boolean
 
     @action
     showSnippetModal() {
@@ -15,20 +15,26 @@ export default class SnippetStore {
     }
 
     @action
-    hideSnippetModal() {
-        this.visible = false
-        if (this.saved) {
+    newSnippet() {
+        if (this.id !== "") {
             this.id = ""
             this.content = ""
-            this.saved = false
         }
+        this.visible = true
+    }
+
+    @action
+    hideSnippetModal() {
+        this.visible = false
     }
 
     @action
     editSnippet(id:string, content:string) {
+        if (this.id !== id) {
+            this.id  = id
+            this.content = content
+        }
         this.visible = true
-        this.id  = id
-        this.content = content
     }
 
     submitSnippet = flow(function * addSnippet(content:string, tags: string[]): any {
@@ -36,6 +42,7 @@ export default class SnippetStore {
             try {
                 let res = yield invokeRPC("addPageX", content, tags)
                 this.id = res.id
+                rs.resources.unshift(res)
                 notify.info("添加成功")
             } catch(err) {
                 notify.warn("添加失败", "error")
@@ -44,12 +51,18 @@ export default class SnippetStore {
         } else {
             try {
                 yield invokeRPC("updatePageX", this.id, content, tags)
-                notify.info("添加成功")
+                notify.info("更新成功")
+                let target = _.find(rs.resources, (o) => {return o.id === this.id})
+                target.content = content
             } catch(err) {
                 notify.warn("添加失败", "error")
                 console.log(err)
             }
         }
-        this.saved = true
     })
 }
+
+let store = new SnippetStore()
+export {store}
+let ctx = createContext(store)
+export default function () { return useContext(ctx)}